# [MRG] MAINT Use numpy vectorized operations in k-means++