Just a small refactoring of k-means++. Better use of numpy instead of looping to find an argmin.

Bonus: a significant speedup (more clusters = more speedup)
```python
import numpy as np 
from sklearn.cluster.k_means_ import _k_init 
X = np.random.RandomState(0).random_sample((10000, 100)) 
xx = (X**2).sum(axis=1) 
%timeit _k_init(X, 100, xx, np.random.RandomState(0)) 
```
On master:
189 ms ± 615 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
With this PR:
159 ms ± 1.53 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)